Social_Listening_Bot/requirement.txt
fastapi
httpx
uvicorn
nltk
python-dotenv
python-multipart
google-cloud-dialogflow-cx